Cleaning

+++ src/size.c
@@ -6,7 +6,7 @@
 /*   By: jterrazz <jterrazz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/09 17:13:09 by jterrazz          #+#    #+#             */
-/*   Updated: 2019/05/22 22:01:50 by jterrazz         ###   ########.fr       */
+/*   Updated: 2019/05/31 18:22:05 by jterrazz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 t_range_group get_range_group_from_block_size(size_t size)
 {
-    if (size <= TINY_BLOCK_SIZE)
+    if (size <= (size_t) TINY_BLOCK_SIZE)
         return (TINY);
-    else if (size <= SMALL_BLOCK_SIZE)
+    else if (size <= (size_t) SMALL_BLOCK_SIZE)
         return (SMALL);
 
     return (LARGE);
@@ -28,9 +28,9 @@ size_t get_range_size_from_block_size(size_t size)
     t_range_group range_group = get_range_group_from_block_size(size);
 
     if (range_group == TINY)
-        return (TINY_RANGE_ALLOCATION_SIZE);
+        return ((size_t) TINY_RANGE_ALLOCATION_SIZE);
     else if (range_group == SMALL)
-        return (SMALL_RANGE_ALLOCATION_SIZE);
+        return ((size_t) SMALL_RANGE_ALLOCATION_SIZE);
 
     return (size + sizeof(t_range) + sizeof(t_block));
 }
